### 2017-03-07 #197 @whedon
[PRE REVIEW]: `learningCurve`: An implementation of Crawford's and Wright's learning curve production functions
**Submitting author:** @bradleyboehmke (<a href="http://orcid.org/0000-0002-3611-8516">Bradley Boehmke</a>)
**Repository:** <a href="https://github.com/Auburngrads/learningCurve" target ="_blank">https://github.com/Auburngrads/learningCurve</a>
**Version:** v1.1.1
