### 2017-03-07 #197 @bradleyboehmke
@arfon , not sure if you'll be the editor on this submission but I know last time you had asked for potential reviewers that were familiar with the mathematical/statistical area that our R package focused on.  For this submission, [Trevor Bihl](https://github.com/TJBihl) (trevor.bihl@gmail.com) would be a good candidate for review.
