### 2017-03-08 #197 @arfon
:wave: @TJBihl - would you be willing to review this submission for JOSS? You can read more about what that involves here: http://joss.theoj.org/about#reviewer_guidelines
