### 2017-03-09 #197 @arfon
@TJBihl - please carry out the review over in #202.
