### 2017-04-20 #202 @TJBihl
The authors have provided a well composed package for learning rates.  I was able to use all of the files/functions with no issues and reproduce all examples.  Beyond this, I was able to explore additional operating points in the functions and saw no issues to correct.

A few minor issues does exist though:
1.  Equations do not appear in the vignettes.  After much searching, this is probably due to the computer I'm using not having a LaTex program installed.  It would help if a comment/disclaimer was included in the install stating that one was needed.  I am using RStudio 1.0.136.  I could possibly need rmarkdown, but this wasn't in the specified library list.
2. paper.md might need the authors names/affiliations
3.  When I first tried running setup.R, it was after I loaded the learning rate package.  But I didn't have the knitr package, so this might be a library the authors should have load by default.
4. paper.bib has some primary references, but the vignette appears to have more.
