### 2017-04-30 #202 @bradleyboehmke
@arfon  How would you like me to proceed?  As for the reviewers comments # 1 & 3, we have included rmarkdown and knitr as suggested dependencies rather than absolute required dependencies.  This is because the package will operate without these packages; however, having them is suggested as it allows for enhancements such as seeing the formulas in the vignette.

Comment #2 - I updated the paper.md yaml so that the authors names/affiliations populate.

Comment #3 - Not sure how you want me to address this.  The vignette has a lot more references because it's written in a full paper format versus JOSS's requirements of keeping it to 1 page.  If I incorporate all the references then I would likely convert the JOSS paper into a much longer full-length paper.

Let me know how to proceed.  Thanks!  And thanks @TJBihl for the solid review!

Brad
