### 2017-04-30 #202 @arfon
> Comment #2 - I updated the paper.md yaml so that the authors names/affiliations populate.

👍 great.

> Comment #3 - Not sure how you want me to address this. The vignette has a lot more references because it's written in a full paper format versus JOSS's requirements of keeping it to 1 page. If I incorporate all the references then I would likely convert the JOSS paper into a much longer full-length paper.

I think it's OK if the JOSS `paper.md` only has a subset of these references (i.e. the ones necessary for the short paper).

@TJBihl - could you verify the changes and check off of the remaining review checkboxes if everything looks good?
