### 2017-05-01 #202 @bradleyboehmke
@arfon as requested:

<a href="https://zenodo.org/badge/latestdoi/60852792"><img src="https://zenodo.org/badge/60852792.svg" alt="DOI"></a>
